# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aarde API",
    "version": "dev",
    "description": "Read-only imagery catalog API served by aarde. Imports are local CLI operations and are not exposed here.\n\nUse **Authorize** to enter the web token from the private token file named in the server startup log. Enter only the token, without the Bearer prefix. Swagger stores authorization in memory and clears it on reload.\n\nAll operations require authentication by default. With AARDE_WEB_PUBLIC_READ=true, registered reads including POST /api/v1/imagery/search accept anonymous requests; this served specification reflects that setting. Explicitly supplied invalid tokens are always rejected. API responses are not cached.\n\nThis reference and /openapi.json are public and contain no catalog data or credentials. Try it out sends real requests to this server.\n\nViewer manifest and layer content always require bearer authentication, even when catalog browsing is public."
  },
  "servers": [
    {
      "url": "/",
      "description": "This aarde server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/imagery": {
      "get": {
        "operationId": "listImagery",
        "summary": "List or filter imagery",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "description": "Exact, case-sensitive IDs. Results are ordered by import time descending, then UUID. Concurrent imports can shift offset-based pages.",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact catalog ID. Omit or use an empty string for all catalogs."
          },
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255,
                "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
              },
              "maxItems": 100
            },
            "description": "Repeat this query parameter for up to 100 exact image IDs.",
            "style": "form",
            "explode": true
          },
          {
            "name": "cloud_cover_lt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover strictly less than this finite percentage (0\u2013100). Zero matches no known values; 100 excludes fully cloudy scenes. Cannot be combined with cloud_cover_lte. Unknown values are excluded by default when a threshold is supplied; cloud_cover_unknown overrides this. All filters combine with AND before ordering and pagination. Omit for no threshold. An empty query value is invalid.",
            "example": 20
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Page size; values outside 1\u2013200 are rejected."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            },
            "description": "Number of records to skip."
          },
          {
            "name": "cloud_cover_lte",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover at most this finite percentage (0\u2013100). Includes the exact boundary, including 0 and 100. Cannot be combined with cloud_cover_lt. Unknown values are excluded by default; use cloud_cover_unknown=include to retain them. Describes the whole scene, not clouds within the search area. An empty query value is invalid."
          },
          {
            "name": "cloud_cover_unknown",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "exclude",
                "include",
                "only"
              ]
            },
            "description": "Unknown cloud-cover policy. Defaults to exclude with either threshold and include otherwise. exclude without a threshold returns only known values; only returns unknown values and cannot be combined with either threshold. An empty query value is invalid."
          },
          {
            "name": "acquired_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive acquisition timestamp (RFC3339). A date bound excludes imagery with unknown acquisition time. Must precede acquired_before when both are supplied. An empty query value is invalid."
          },
          {
            "name": "acquired_before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Exclusive acquisition timestamp (RFC3339). For an inclusive Through day, send midnight UTC of the following day. A date bound excludes imagery with unknown acquisition time. An empty query value is invalid."
          }
        ]
      },
      "head": {
        "operationId": "listImageryHead",
        "summary": "List or filter imagery (headers only)",
        "tags": [
          "Catalog"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "A valid aarde web token is required"
          },
          "400": {
            "description": "Invalid pagination, IDs, date bounds, cloud thresholds, or unknown-cloud policy."
          },
          "500": {
            "description": "catalog request failed"
          }
        },
        "description": "Exact, case-sensitive IDs. Results are ordered by import time descending, then UUID. Concurrent imports can shift offset-based pages.",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact catalog ID. Omit or use an empty string for all catalogs."
          },
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255,
                "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
              },
              "maxItems": 100
            },
            "description": "Repeat this query parameter for up to 100 exact image IDs.",
            "style": "form",
            "explode": true
          },
          {
            "name": "cloud_cover_lt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover strictly less than this finite percentage (0\u2013100). Zero matches no known values; 100 excludes fully cloudy scenes. Cannot be combined with cloud_cover_lte. Unknown values are excluded by default when a threshold is supplied; cloud_cover_unknown overrides this. All filters combine with AND before ordering and pagination. Omit for no threshold. An empty query value is invalid.",
            "example": 20
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Page size; values outside 1\u2013200 are rejected."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            },
            "description": "Number of records to skip."
          },
          {
            "name": "cloud_cover_lte",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100
            },
            "description": "Reported scene cloud cover at most this finite percentage (0\u2013100). Includes the exact boundary, including 0 and 100. Cannot be combined with cloud_cover_lt. Unknown values are excluded by default; use cloud_cover_unknown=include to retain them. Describes the whole scene, not clouds within the search area. An empty query value is invalid."
          },
          {
            "name": "cloud_cover_unknown",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "exclude",
                "include",
                "only"
              ]
            },
            "description": "Unknown cloud-cover policy. Defaults to exclude with either threshold and include otherwise. exclude without a threshold returns only known values; only returns unknown values and cannot be combined with either threshold. An empty query value is invalid."
          },
          {
            "name": "acquired_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive acquisition timestamp (RFC3339). A date bound excludes imagery with unknown acquisition time. Must precede acquired_before when both are supplied. An empty query value is invalid."
          },
          {
            "name": "acquired_before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Exclusive acquisition timestamp (RFC3339). For an inclusive Through day, send midnight UTC of the following day. A date bound excludes imagery with unknown acquisition time. An empty query value is invalid."
          }
        ]
      }
    },
...
    "/api/v1/imagery/{catalogID}/{imageID}/viewer": {
      "get": {
        "operationId": "getImageViewer",
        "summary": "Get the complete physical-file viewer manifest",
        "tags": [
          "Viewer"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "catalogID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            },
            "description": "Catalog identifier."
          },
          {
            "name": "imageID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
            },
            "description": "Image identifier within the catalog."
          },
          {
            "name": "resolution",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "auto",
                "preview",
                "native"
              ],
              "default": "auto"
            },
            "description": "Whole-image display budget. Native is rejected if any scene/layer limit is exceeded."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerManifest"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifiers or rendering parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Catalog source or layer unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Request cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Source changed; reload manifest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported or failed decoding/registration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Whole-image resolution exceeds configured limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Catalog failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Render capacity exhausted; retry later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Request time limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Always authenticated. One complete segment image per raster response. Placement mesh maps original pixel edges into a common image-coordinate plane (x right, y up). Display downsampling never changes placement. Geometry coordinates use this same manifest coordinate system, not RFC7946 longitude/latitude. Opacity/pan/zoom do not request content. Cloud grids support NCDRD 2010 Byte values 0/255 with reserved values shown distinctly; CSSHPA v01 polygons are bounded and transformed. No RPC/DEM orthorectification."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Startup web token. Enter the token only, without the Bearer prefix."
      }
    },
    "schemas": {
...
      "Position": {
        "type": "array",
        "prefixItems": [
          {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          }
        ],
        "items": false,
        "minItems": 2,
        "maxItems": 2
      },
...
      "MultiPolygon": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "const": "MultiPolygon"
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Position"
                },
                "minItems": 4
              },
              "minItems": 1
            },
            "minItems": 1
          }
        },
        "required": [
          "type",
          "coordinates"
        ],
        "additionalProperties": false,
        "example": {
          "type": "MultiPolygon",
          "coordinates": [
            [
              [
                [
                  -106.2,
                  39.3
                ],
                [
                  -105.8,
                  39.3
                ],
                [
                  -105.8,
                  39.7
                ],
                [
                  -106.2,
                  39.7
                ],
                [
                  -106.2,
                  39.3
                ]
              ]
            ]
          ]
        }
      },
...
      "Imagery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "catalog_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
          },
          "image_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Exact, case-sensitive ID: 1\u2013255 UTF-8 bytes, no slashes, backslashes, control characters, or surrounding whitespace."
          },
          "display_name": {
            "type": "string"
          },
          "acquired_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "For multi-image NITF, a unanimous known segment time, otherwise null."
          },
          "imported_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "footprint": {
            "$ref": "#/components/schemas/MultiPolygon"
          },
          "checksum": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "description": "SHA-256 of the source file."
          },
          "asset_location": {
            "type": "string",
            "description": "Local asset path on the server; this API does not download raster files."
          },
          "width": {
            "type": "integer",
            "minimum": 0,
            "description": "Positive for a single raster; 0 (unavailable) for multi-image NITF. See segments."
          },
          "height": {
            "type": "integer",
            "minimum": 0,
            "description": "Positive for a single raster; 0 (unavailable) for multi-image NITF. See segments."
          },
          "band_count": {
            "type": "integer",
            "minimum": 0,
            "description": "Positive for a single raster; 0 (unavailable) for multi-image NITF. See segments."
          },
          "source_crs": {
            "type": "string",
            "description": "Source CRS for a single raster; empty for multi-image NITF. See segments."
          },
          "metadata": {
            "description": "Selected GDAL metadata domains preserved as JSON, including object and XML values. The reserved _aarde namespace contains format, georeferencing, and footprint-method annotations; handling markings are metadata, not access controls."
          },
          "cloud_cover": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 100,
            "description": "File-level override, or unanimous known segment percentage for multi-image NITF, otherwise null."
          },
          "format": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "GTiff",
              "NITF",
              null
            ],
            "description": "GDAL-detected source format, stored at import. Null for records imported before format annotations; never inferred from the filename."
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageSegment"
            },
            "description": "NITF image segments in zero-based index order; absent or empty for legacy records and GeoTIFF."
          }
        },
        "required": [
          "id",
          "catalog_id",
          "image_id",
          "display_name",
          "acquired_at",
          "imported_at",
          "created_at",
          "footprint",
          "checksum",
          "asset_location",
          "width",
          "height",
          "band_count",
          "source_crs",
          "metadata",
          "cloud_cover",
          "format"
        ],
        "example": {
          "id": "c77f4d82-ef3d-44b9-b027-c6fc30b60c53",
          "catalog_id": "default",
          "image_id": "ABC123",
          "display_name": "ABC123.tif",
          "acquired_at": null,
          "imported_at": "2026-09-27T12:00:00Z",
          "created_at": "2026-09-27T12:00:00Z",
          "footprint": {
            "type": "MultiPolygon",
            "coordinates": [
              [
                [
                  [
                    -106.2,
                    39.3
                  ],
                  [
                    -105.8,
                    39.3
                  ],
                  [
                    -105.8,
                    39.7
                  ],
                  [
                    -106.2,
                    39.7
                  ],
                  [
                    -106.2,
                    39.3
                  ]
                ]
              ]
            ]
          },
          "checksum": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
          "asset_location": "/data/ABC123.tif",
          "width": 1024,
          "height": 1024,
          "band_count": 3,
          "source_crs": "GEOGCS[\"WGS 84\",DATUM[\"WGS_1984\",SPHEROID[\"WGS 84\",6378137,298.257223563]],PRIMEM[\"Greenwich\",0],UNIT[\"degree\",0.0174532925199433]]",
          "metadata": {},
          "cloud_cover": null,
          "format": null
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Imagery"
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "default": 50
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000,
            "default": 0
          },
          "has_more": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "example": {
          "items": [],
          "limit": 50,
          "offset": 0,
          "has_more": false
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": {
            "code": "invalid_geometry",
            "message": "search geometry must be a Polygon or MultiPolygon"
          }
        }
      },
...
      "ViewerManifest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "catalog_id": {
            "type": "string"
          },
          "image_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "coordinate_system": {
            "type": "string"
          },
          "resolution": {
            "type": "string"
          },
          "cloud_status": {
            "type": "string"
          },
          "extent": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "minItems": 4,
            "maxItems": 4
          },
          "layers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ViewerLayer"
            }
          },
          "native_available": {
            "type": "boolean"
          },
          "display_pixels": {
            "type": "integer",
            "minimum": 0
          },
          "estimated_memory_bytes": {
            "type": "integer",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ncdrd_assessment": {
            "type": "object"
          }
        },
        "required": [
          "catalog_id",
          "image_id",
          "name",
          "revision",
          "format",
          "coordinate_system",
          "resolution",
          "cloud_status",
          "extent",
          "layers",
          "native_available",
          "display_pixels",
          "estimated_memory_bytes",
          "warnings"
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "A valid aarde web token is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "A valid aarde web token is required"
              }
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "example": "Bearer realm=\"aarde-web\""
          }
        }
      },
      "InvalidRequest": {
        "description": "Invalid pagination, catalog/image IDs, RFC3339 date bounds, cloud thresholds, or cloud_cover_unknown policy. Thresholds must be finite percentages from 0 to 100 and cannot be combined. Unknown-only cannot be combined with a threshold. Spatial search requires valid polygon geometry and a single JSON object with known fields.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "invalid_request",
                "message": "invalid limit"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "catalog request failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "catalog request failed"
              }
            }
          }
        }
      },
...
    }
  }
}